Trait chip8::chip8::keyboard::Keyboard

pub trait Keyboard {
    fn is_key_pressed(&self, key: u8) -> bool;

    fn wait_for_key_press(&self) -> u8;

    fn is_quit(&self) -> bool;
}

A trait that encapsulates the functionality required for keyboard IO. This trait exists to let us mock the keyboard in unit tests The chip-8 keyboard has 16 hexadecimal keys.

Required methods

fn is_key_pressed(&self, key: u8) -> bool

fn wait_for_key_press(&self) -> u8

fn is_quit(&self) -> bool

Loading content...